changed weight to double and rounded to two decimal points

+++ src/edge.h
@@ -67,7 +67,7 @@ class Edge
     /**
      * Gets edge weight.
      */
-    int getDistance() const
+    double getDistance() const
     {
         return this->distance;
     }
